requirement.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
PyMuPDF==1.23.8
python-docx==1.1.0
nltk==3.8.1
langdetect==1.0.9
scikit-learn==1.3.2
numpy==1.26.2
beautifulsoup4==4.12.2
requests==2.31.0
plotly==5.18.0
reportlab==4.0.7
werkzeug==3.0.1
chardet==5.2.0